BTAP Results Test Suite: Moving the test helper to the correct directory and fixing includes
