+++ test/btap_costing/tests/BtapResults_HS09_CCASHP_Baseboard_Test.rb
@@ -1,12 +1,12 @@
-require_relative '../../../lib/openstudio-standards.rb'
-require_relative '../../../lib/openstudio-standards/btap/btap_test_helper.rb'
+require_relative '../../../lib/openstudio-standards'
+require_relative '../../helpers/btap_results_helper'
 require 'minitest/autorun'
 require 'optparse'
 require 'fileutils'
 require 'minitest/unit'
 require 'optparse'
 
-class BTAPResults_HS09_Test < Minitest::Test
+class BTAP::Results_HS09_Test < Minitest::Test
   def test_results_hs09()
     #building_type = 'Outpatient'
     #building_type = 'LargeHotel'
@@ -91,7 +91,7 @@ class BTAPResults_HS09_Test < Minitest::Test
                                                    furnace_eff: furnace_eff,
                                                    unitary_cop: unitary_cop,
                                                    shw_eff: shw_eff,
-                                                   cached: BTAPResultsHelper.cached)
+                                                   cached: BTAP::ResultsHelper.cached)
   end
 
   def create_model_simulate_and_qaqc_regression_test(epw_file:,
@@ -113,7 +113,7 @@ class BTAPResults_HS09_Test < Minitest::Test
     model_name = "#{building_type}-#{template}-DefaultFuel-#{File.basename(epw_file, '.epw')}_CCASHP"
     test_dir   = "#{File.dirname(__FILE__)}/output"
     run_dir    = "#{test_dir}/#{model_name}"
-    helper     = BTAPResultsHelper.new(test_path: __FILE__, model_name: model_name, run_dir: run_dir)
+    helper     = BTAP::ResultsHelper.new(test_path: __FILE__, model_name: model_name, run_dir: run_dir)
 
     Dir.mkdir(test_dir) unless Dir.exist?(test_dir)
     Dir.mkdir(run_dir) unless Dir.exist?(run_dir)
