+++ test/btap_costing/tests/BtapResults_SHW_Test.rb
@@ -1,12 +1,12 @@
-require_relative '../../../lib/openstudio-standards.rb'
-require_relative '../../../lib/openstudio-standards/btap/btap_test_helper.rb'
+require_relative '../../../lib/openstudio-standards'
+require_relative '../../helpers/btap_results_helper'
 require 'minitest/autorun'
 require 'optparse'
 require 'fileutils'
 require 'minitest/unit'
 require 'optparse'
 
-class BTAPResults_SHW_Test < Minitest::Test
+class BTAP::Results_SHW_Test < Minitest::Test
   def test_results_shw()
     #building_type = 'Outpatient'
     #building_type = 'LargeHotel'
@@ -91,7 +91,7 @@ class BTAPResults_SHW_Test < Minitest::Test
                                                    furnace_eff: furnace_eff,
                                                    unitary_cop: unitary_cop,
                                                    shw_eff: shw_eff,
-                                                   cached: BTAPResultsHelper.cached)
+                                                   cached: BTAP::ResultsHelper.cached)
 
 
   end
@@ -115,7 +115,7 @@ class BTAPResults_SHW_Test < Minitest::Test
     model_name = "#{building_type}-#{template}-DefaultFuel-#{File.basename(epw_file, '.epw')}_SHW"
     test_dir   = "#{File.dirname(__FILE__)}/output"
     run_dir    = "#{test_dir}/#{model_name}"
-    helper     = BTAPResultsHelper.new(test_path: __FILE__, model_name: model_name, run_dir: run_dir)
+    helper     = BTAP::ResultsHelper.new(test_path: __FILE__, model_name: model_name, run_dir: run_dir)
 
     Dir.mkdir(test_dir) unless Dir.exist?(test_dir)
     Dir.mkdir(run_dir) unless Dir.exist?(run_dir)
