+++ test/btap_costing/tests/BtapResults_insuite_central_doas_Test.rb
@@ -1,12 +1,12 @@
-require_relative '../../../lib/openstudio-standards.rb'
-require_relative '../../../lib/openstudio-standards/btap/btap_test_helper.rb'
+require_relative '../../../lib/openstudio-standards'
+require_relative '../../helpers/btap_results_helper'
 require 'minitest/autorun'
 require 'optparse'
 require 'fileutils'
 require 'minitest/unit'
 require 'optparse'
 
-class BTAPResults_DOAS_Test < Minitest::Test
+class BTAP::Results_DOAS_Test < Minitest::Test
   def test_results_doas()
     #building_type = 'Outpatient'
     #building_type = 'LargeHotel'
@@ -38,7 +38,7 @@ class BTAPResults_DOAS_Test < Minitest::Test
                                                    template: template,
                                                    building_type: building_type,
                                                    baseline_system_zones_map_option: baseline_system_zones_map_option,
-                                                   cached: BTAPResultsHelper.cached)
+                                                   cached: BTAP::ResultsHelper.cached)
   end
 
   def create_model_simulate_and_qaqc_regression_test(epw_file:,
@@ -50,7 +50,7 @@ class BTAPResults_DOAS_Test < Minitest::Test
     model_name = "#{building_type}-#{template}-#{File.basename(epw_file, '.epw')}-doas-type-#{baseline_system_zones_map_option}"
     test_dir   = "#{File.dirname(__FILE__)}/output"
     run_dir    = "#{test_dir}/#{model_name}"
-    helper     = BTAPResultsHelper.new(test_path: __FILE__, model_name: model_name, run_dir: run_dir)
+    helper     = BTAP::ResultsHelper.new(test_path: __FILE__, model_name: model_name, run_dir: run_dir)
 
     Dir.mkdir(test_dir) unless Dir.exist?(test_dir)
     Dir.mkdir(run_dir) unless Dir.exist?(run_dir)
