+++ test/btap_costing/tests/BtapResults_led_lighting_Test.rb
@@ -1,12 +1,12 @@
-require_relative '../../../lib/openstudio-standards.rb'
-require_relative '../../../lib/openstudio-standards/btap/btap_test_helper.rb'
+require_relative '../../../lib/openstudio-standards'
+require_relative '../../helpers/btap_results_helper'
 require 'minitest/autorun'
 require 'optparse'
 require 'fileutils'
 require 'minitest/unit'
 require 'optparse'
 
-class BTAPResults_LED_Test < Minitest::Test
+class BTAP::Results_LED_Test < Minitest::Test
   def test_results_led()
     #building_type = 'Outpatient'
     #building_type = 'LargeHotel'
@@ -41,7 +41,7 @@ class BTAPResults_LED_Test < Minitest::Test
                                                    building_type: building_type,
                                                    lights_type: lights_type,
                                                    lights_scale: lights_scale,
-                                                   cached: BTAPResultsHelper.cached)
+                                                   cached: BTAP::ResultsHelper.cached)
   end
 
   def create_model_simulate_and_qaqc_regression_test(epw_file:,
@@ -54,7 +54,7 @@ class BTAPResults_LED_Test < Minitest::Test
     model_name = "#{building_type}-#{template}-#{File.basename(epw_file, '.epw')}-#{lights_type}"
     test_dir   = "#{File.dirname(__FILE__)}/output"
     run_dir    = "#{test_dir}/#{model_name}"
-    helper     = BTAPResultsHelper.new(test_path: __FILE__, model_name: model_name, run_dir: run_dir)
+    helper     = BTAP::ResultsHelper.new(test_path: __FILE__, model_name: model_name, run_dir: run_dir)
 
     Dir.mkdir(test_dir) unless Dir.exist?(test_dir)
     Dir.mkdir(run_dir) unless Dir.exist?(run_dir)
