+++ test/btap_costing/tests/BtapResults_pv_ground_Test.rb
@@ -1,12 +1,12 @@
-require_relative '../../../lib/openstudio-standards.rb'
-require_relative '../../../lib/openstudio-standards/btap/btap_test_helper.rb'
+require_relative '../../../lib/openstudio-standards'
+require_relative '../../helpers/btap_results_helper'
 require 'minitest/autorun'
 require 'optparse'
 require 'fileutils'
 require 'minitest/unit'
 require 'optparse'
 
-class BTAPResults_PV_Test < Minitest::Test
+class BTAP::Results_PV_Test < Minitest::Test
   def test_results_pv()
     #building_type = 'Outpatient'
     #building_type = 'LargeHotel'
@@ -45,7 +45,7 @@ class BTAPResults_PV_Test < Minitest::Test
                                                    pv_ground_tilt_angle: pv_ground_tilt_angle,
                                                    pv_ground_azimuth_angle: pv_ground_azimuth_angle,
                                                    pv_ground_module_description: pv_ground_module_description,
-                                                   cached: BTAPResultsHelper.cached)
+                                                   cached: BTAP::ResultsHelper.cached)
   end
 
   def create_model_simulate_and_qaqc_regression_test(epw_file:,
@@ -61,7 +61,7 @@ class BTAPResults_PV_Test < Minitest::Test
     model_name = "#{building_type}-#{template}-#{File.basename(epw_file, '.epw')}-pv-ground-type-#{true}"
     test_dir   = "#{File.dirname(__FILE__)}/output"
     run_dir    = "#{test_dir}/#{model_name}"
-    helper     = BTAPResultsHelper.new(test_path: __FILE__, model_name: model_name, run_dir: run_dir)
+    helper     = BTAP::ResultsHelper.new(test_path: __FILE__, model_name: model_name, run_dir: run_dir)
 
     Dir.mkdir(test_dir) unless Dir.exist?(test_dir)
     Dir.mkdir(run_dir) unless Dir.exist?(run_dir)
